feat: added new entity player
Working on getting the engine running. The current implementation is not flexible
to allow for adding in additional features and there is a lot of duplicate code
that could be better implemented with inheritance from a base object.

This is a mark in time, to capture the current state of the project.